Use libc::Ioctl for ioctl constant
Unfortunately musl uses int and glibc uses unsigned long for the ioctl
request. The actual kernel sees it as a packed 32 bit value.

The libc crate tests that the symbols it exports have the same
signature as the underlying libc. It at least export a Ioctl type that
we can use to compile against both glibc and musl without too much
pain.