Update deps
tar 0.4.36 and newer fail because it will only create directories that
are in the archive. Since we're specifying an additional directory to
extract into, we must create it ourselves.

Several deps bumped MSRV as part of patch releases. These must be
updated after a general update back to versions that do not break
builds.

    cargo update -p idna_adapter --precise 1.2.0
    cargo update -p litemap --precise 0.7.4
    cargo update -p zerofrom --precise 0.1.5

Signed-off-by: Tim Crawford <tcrawford@system76.com>